Refactored amami entry point (main() function) to be inside amami/__init__.py, instead of being inside an additional file amami/amami.py
