+++ setup.cfg
@@ -32,4 +32,4 @@ install_requires =
 
 [entry_points]
 console_scripts = 
-    amami=amami.amami:main
+    amami=amami:main
